backend/requirements.txt
Flask
Flask-CORS
git+https://github.com/openai/whisper.git
openai
pinecone-client
sqlalchemy
psycopg2-binary
python-dotenv
qdrant-client